+++ crates/oxide/src/lib.rs
@@ -447,7 +447,14 @@ fn read_changed_content(c: ChangedContent) -> Option<Vec<u8>> {
     };
 
     match extension {
-        Some("svelte") => Some(content.replace(" class:", " ")),
+        // Angular class shorthand
+        Some("html") => Some(content.replace("[class.", "[")),
+        Some("svelte") => Some(
+            content
+                .replace(" class:", " ")
+                .replace("\tclass:", " ")
+                .replace("\nclass:", " "),
+        ),
         _ => Some(content),
     }
 }