+++ crates/oxide/tests/scanner.rs
@@ -323,14 +323,33 @@ mod scanner {
             ("foo.jpg", "xl:font-bold"),
             // A file that is ignored
             ("foo.html", "lg:font-bold"),
+            // An Angular file using the class shorthand syntax
+            (
+                "index.angular.html",
+                "<div [class.underline]=\"bool\"></div>",
+            ),
             // A svelte file with `class:foo="bar"` syntax
             ("index.svelte", "<div class:px-4='condition'></div>"),
+            ("index2.svelte", "<div\n\tclass:px-5='condition'></div>"),
+            ("index3.svelte", "<div\n  class:px-6='condition'></div>"),
+            ("index4.svelte", "<div\nclass:px-7='condition'></div>"),
         ])
         .1;
 
         assert_eq!(
             candidates,
-            vec!["condition", "div", "font-bold", "md:flex", "px-4"]
+            vec![
+                "bool",
+                "condition",
+                "div",
+                "font-bold",
+                "md:flex",
+                "px-4",
+                "px-5",
+                "px-6",
+                "px-7",
+                "underline"
+            ]
         );
     }